fixed bug where double was not a subtype of int

+++ lib/core/models/overview.model.dart
@@ -18,15 +18,16 @@ class Overview {
   factory Overview.fromJson(Map<String, dynamic> json) => Overview(
         activeSubscribersCount: json["active_subscribers_count"] == null
             ? null
-            : json["active_subscribers_count"],
+            : json["active_subscribers_count"].toDouble(),
         activeTrialsCount: json["active_trials_count"] == null
             ? null
-            : json["active_trials_count"],
+            : json["active_trials_count"].toDouble(),
         activeUsersCount: json["active_users_count"] == null
             ? null
-            : json["active_users_count"],
-        installsCount:
-            json["installs_count"] == null ? null : json["installs_count"],
+            : json["active_users_count"].toDouble(),
+        installsCount: json["installs_count"] == null
+            ? null
+            : json["installs_count"].toDouble(),
         mrr: json["mrr"] == null ? null : json["mrr"].toDouble(),
         revenue: json["revenue"] == null ? null : json["revenue"].toDouble(),
       );
